tests: Move scheduler fixtures to test/conftest
This allows us to use them for integration tests and acceptance test
